requirements.txt
astrapy
python-dotenv
transformers
torch
Pillow